Allow dates to be configured
In some cases dates may not be desired. This adds a simple config switch
that enables/disables them.

Note: if the config param `showPostedDate` is not set at all then dates
won't be shown. This changes the current behaviour. However the example
site defaults `showPostedDate` to `true` so new sites will maintain the
existing behaviour.

tl;dr: sites that upgrade the git repo in existing prices might be
surprised to see their dates disappear.

Of course we could also go with a `hidePostedDate` but I wanted to avoid
inconsistency with the other `show*` variables and double negatives in
the code itself.