Use tailwind typography for written content
Tailwind's typography is a great baseline for written content and, to my
eye, improves readability of the core content.

This adds tailwind as a css file that generates a scss file that's
included in the usual workflow.

It's not an ideal situation, as we have two resets in play at once, and
tailwind's styles need overriding in places. In future I can see moving
away to use tailwind exclusively, but for now this is a solid step.