Allow for custom punctuation
The dot works for many cases, but now that I'm using it on the content
pages themselves it can conflict with titles that end in other
punctuation. I.e. `?` or `;`.

This allows a `punctuation` key in the frontmatter that overrides the
dot. It's used on the content pages and on list pages too.